:truck: Split cores of packages
We are doing this in order to prevent from having to depend on heavy packages
