+++ src/core/Nihil/Syntax/Core.hs
@@ -1,0 +1,2 @@
+module Nihil.Syntax.Core where
+
